Initial refactor push. Moved files around into respective folders and out of common folder. Removed unused components. Imported components from new locations. Also created walkscore info for modal reducing lines of code by a net-40 count. Now I will get a better understanding of our api calls and useeffects on plotly to begin consolidating useState objects and possible add contextAPI as well
