+++ src/api/dsapi.js
@@ -1,4 +1,3 @@
-import { API } from 'aws-amplify';
 import axios from 'axios';
 
 // Base url to make request to ds API
