--- src/components/pages/CompareCities/compCityContainer.js
@@ -1,22 +1,0 @@
-import React from 'react';
-
-import { getProfileData } from '../../../api';
-
-import { List } from '../../common';
-
-// Here is an example of using our reusable List component to display some list data to the UI.
-const ProfileList = () => {
-  return (
-    <List
-      // Here we are passing our Axios request helper function as a callback.
-      getItemsData={() => getProfileData()}
-      // Here we are passing in a component we want to show whilst waiting for our API request
-      // to complete.
-      LoadingComponent={() => <div>Loading Profiles...</div>}
-      // Here we are passing in a component that receives our new data and returns our JSX elements.
-      // RenderItems={RenderProfileListPage}
-    />
-  );
-};
-
-export default ProfileList;
