--- src/components/pages/CompareCities/index.js
@@ -1,1 +1,0 @@
-export { default as ProfileListPage } from './compCityContainer';
