--- src/components/pages/ExampleDataViz/index.js
@@ -1,3 +1,0 @@
-import DataVizContainer from './ExampleDataVizContainer';
-
-export { DataVizContainer as ExampleDataViz };
