--- src/components/pages/ExampleDataViz/statedata.js
@@ -1,55 +1,0 @@
-const states = [
-  { value: 'AL' },
-  { value: 'AK' },
-  { value: 'AR' },
-  { value: 'AZ' },
-  { value: 'CA' },
-  { value: 'CO' },
-  { value: 'CT' },
-  { value: 'DC' },
-  { value: 'DE' },
-  { value: 'FL' },
-  { value: 'GA' },
-  { value: 'HI' },
-  { value: 'IA' },
-  { value: 'ID' },
-  { value: 'IL' },
-  { value: 'IN' },
-  { value: 'KS' },
-  { value: 'KY' },
-  { value: 'LA' },
-  { value: 'MA' },
-  { value: 'MD' },
-  { value: 'ME' },
-  { value: 'MI' },
-  { value: 'MN' },
-  { value: 'MO' },
-  { value: 'MS' },
-  { value: 'MT' },
-  { value: 'NC' },
-  { value: 'NE' },
-  { value: 'NH' },
-  { value: 'NJ' },
-  { value: 'NM' },
-  { value: 'NV' },
-  { value: 'NY' },
-  { value: 'ND' },
-  { value: 'OH' },
-  { value: 'OK' },
-  { value: 'OR' },
-  { value: 'PA' },
-  { value: 'RI' },
-  { value: 'SC' },
-  { value: 'SD' },
-  { value: 'TN' },
-  { value: 'TX' },
-  { value: 'UT' },
-  { value: 'VT' },
-  { value: 'VA' },
-  { value: 'WA' },
-  { value: 'WI' },
-  { value: 'WV' },
-  { value: 'WY' },
-];
-
-export default states;
